requirements.txt
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.1
Pillow==10.3.0
plotly==5.18.0
streamlit==1.30.0
streamlit_option_menu==0.3.12
